refactor(repository): isolate repo authority tests
